backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.129.0
greenlet==3.1.1
groq==0.37.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
langchain-core==1.2.16
langchain-groq==1.1.2
langgraph==1.0.10
lxml==5.3.1
numpy==2.2.3
openai==1.69.0
opencv-python-headless==4.11.0.86
passlib==1.7.4
Pillow==11.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.40.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
